Moved Linux platform directory to "generic_unix", and test for platform in CMake files
