+++ src/CMakeLists.txt
@@ -9,7 +9,15 @@ endif()
 
 include_directories(${CMAKE_CURRENT_BINARY_DIR} libAtomVM/)
 
-add_executable(AtomVM main.c platforms/${CMAKE_SYSTEM_NAME}/sys.c platforms/${CMAKE_SYSTEM_NAME}/mapped_file.c platforms/${CMAKE_SYSTEM_NAME}/udpdriver.c)
+if(${CMAKE_SYSTEM_NAME} STREQUAL "Darwin")
+    set(PLATFORM_DIR "generic_unix")
+elseif(${CMAKE_SYSTEM_NAME} STREQUAL "Linux")
+    set(PLATFORM_DIR "generic_unix")
+else()
+    message(FATAL_ERROR "Unsupported platform: ${CMAKE_SYSTEM_NAME}")
+endif()
+
+add_executable(AtomVM main.c platforms/${PLATFORM_DIR}/sys.c platforms/${PLATFORM_DIR}/mapped_file.c platforms/${PLATFORM_DIR}/udpdriver.c)
 target_link_libraries(AtomVM libAtomVM)
 set_property(TARGET AtomVM PROPERTY C_STANDARD 99)
